code reorganized, one import at time on server, fix select all when imported items
