+++ index.js
@@ -1,7 +1,3 @@
-const fs = require('fs');
-const { IncomingMessage } = require('http');
-const unlink = require('util').promisify(fs.unlink);
-
 module.exports = {
   improve: 'apostrophe-images',
   moogBundle: {
@@ -11,236 +7,15 @@ module.exports = {
       'apostrophe-media-sources-wedia'
     ]
   },
-  construct: function(self, options) {
-    const superDefineCursor = self.defineCursor;
-    self.defineCursor = () => {
-      superDefineCursor(self.defineCursor);
-      self.apos.define('apostrophe-images-cursor', require('./lib/improveCursor.js'));
-    };
-
-    self.pushAsset('script', 'improve-manager-modal', {
-      when: 'user',
-      preshrunk: true
-    });
-
-    self.on('apostrophe:modulesReady', 'getAllImagesConnectorsModules', () => {
-      // Find all images connectors defined in app configuration
-      self.connectors = Object.values(self.apos.modules)
-        .reduce((connectors, { options: moduleOptions }) => {
-          return !moduleOptions.mediaSourceConnector
-            ? connectors
-            : [
-              ...connectors,
-              {
-                name: moduleOptions.name,
-                label: moduleOptions.label,
-                action: `${self.action}`,
-                ...moduleOptions.mediaSourceConnector
-              }
-            ];
-        }, []);
-
-      const mediaSourcesFiltersOptions = self.connectors.map(({ name, label }) => {
-        return {
-          value: name,
-          label
-        };
-      });
-
-      if (mediaSourcesFiltersOptions.length) {
-        self.filters.push({
-          name: 'mediaSource',
-          choices: [
-            {
-              value: 'apostrophe',
-              label: 'Apostrophe'
-            },
-            ...mediaSourcesFiltersOptions
-          ],
-          def: null
-        });
-      }
-
-      self.on('apostrophe-pages:beforeSend', 'sendConnectorsToBrowser', (req) => {
-        req.browserCall('apos.mediaSourceConnectors=?', JSON.stringify(self.connectors));
-      });
-    });
-
-    self.route('post', 'media-sources-browser', function(req, res) {
-      const { provider } = req.body;
-
-      const mediaSourceConnector = self.connectors
-        .find((connector) => connector.label === provider);
-
-      const choices = self.apos.modules[mediaSourceConnector.name].choices();
-
-      const standardFilters = (mediaSourceConnector.standardFilters || [])
-        .reduce((acc, filter) => {
-          return {
-            ...acc,
-            [filter.name]: filter
-          };
-        }, {});
-
-      return self.renderAndSend(req, 'mediaSourcesBrowser', {
-        ...mediaSourceConnector,
-        standardFilters,
-        choices
-      });
-    });
-
-    self.route('post', 'media-sources-preview', function(req, res) {
-      const {
-        provider, item, isImported
-      } = req.body;
-
-      const mediaSourceConnector = self.connectors
-        .find((connector) => connector.label === provider);
-
-      return self.renderAndSend(req, 'mediaSourcesPreview', {
-        ...mediaSourceConnector,
-        item,
-        isImported
-      });
-    });
-
-    self.route('post', 'find', async function(req, res) {
-      try {
-        const { connector, ...filters } = req.body;
-        const currentModule = self.apos.modules[connector];
-
-        if (self.isConnectorWithMethod(currentModule, 'find')) {
-          const images = await currentModule.find(req, filters);
-
-          const existingImages = await self.apos.images.find(req, {
-            mediaSource: connector,
-            mediaSourceId: { $in: images.results.map((img) => img.mediaSourceId) }
-          }).toArray();
-
-          const data = {
-            images,
-            existingIds: existingImages.map((img) => img.mediaSourceId),
-            filterChoices: currentModule.choices(req, filters)
-          };
-
-          return res.status(200).send(data);
-        }
-
-        res.status(404).send(`Connector ${connector} doesn't exist or doesn't have the right methods`);
-      } catch (err) {
-        if (err.response) {
-          const { status, data } = err.response;
-          return res.status(status || 500).send(data);
-        }
-
-        res.status(500).send(err);
-      }
-    });
-
-    self.route('post', 'download', async function(req, res) {
-      try {
-        const { connector, files } = req.body;
-        const currentModule = self.apos.modules[connector];
-
-        if (!currentModule) {
-          return res.status(400).send(`This module doesn't exist: ${connector}.`);
-        }
-
-        const data = await self.downloadAndInsert(req, files, currentModule);
-
-        if (!data) {
-          throw new Error(`Error while downloading image for ${connector}.`);
-        }
-
-        return res.status(200).send(data);
-      } catch (err) {
-        if (err.response) {
-          const {
-            status, statusText, data
-          } = err.response;
-          const dataIsStream = data instanceof IncomingMessage;
-
-          return res.status(status || 500).send(dataIsStream ? statusText : data);
-        }
-
-        res.status(500).send(err);
-      }
-    });
-
-    self.isConnectorWithMethod = (module, method) => {
-      return module && module[method] &&
-        typeof module[method] === 'function' &&
-        module.options.mediaSourceConnector;
-    };
-
-    self.downloadAndInsert = async (req, files, connectorModule) => {
-      const imagesIds = [];
-      const tempPaths = [];
-
-      const tempPath = self.apos.attachments.uploadfs.getTempPath();
-
-      for (const file of files) {
-        // If a script is used, we need to format the file in the download method
-        // so we have to return it to insert image with the right file object
-        const { fileName, formattedFile } = await connectorModule
-          .download(req, file, tempPath);
-
-        const fileToInsert = formattedFile || file;
-
-        const filePath = `${tempPath}/${fileName}`;
-
-        const image = await self.insertImage({
-          req,
-          connectorName: connectorModule.options.name,
-          file: fileToInsert,
-          fileName,
-          filePath
-        });
-
-        tempPaths.push(filePath);
-        imagesIds.push(image._id);
-      }
-
-      self.unlinkImages(tempPaths);
-
-      return imagesIds;
-    };
-
-    self.insertImage = async ({
-      req,
-      connectorName,
-      file,
-      fileName,
-      filePath
-    }) => {
-      const { slugPrefix } = self.apos.modules['apostrophe-images'].options;
-
-      const attachment = await self.apos.attachments.insert(req, {
-        name: fileName,
-        path: filePath
-      });
-
-      const title = file.title || file.mediaSourceId;
-
-      const piece = {
-        title,
-        slug: `${slugPrefix || ''}${file.mediaSourceId}`,
-        published: true,
-        trash: false,
-        attachment,
-        mediaSourceId: file.mediaSourceId,
-        mediaSource: connectorName
-      };
-
-      return self.apos.images.insert(req, piece);
-    };
-
-    self.unlinkImages = (tempPaths) => {
-      const promises = tempPaths.map((tempPath) => {
-        return unlink(tempPath);
-      });
-
-      return Promise.all(promises);
-    };
+  construct: (self, options) => {
+    require('./lib/api.js')(self, options);
+    require('./lib/routes.js')(self, options);
+    require('./lib/assets.js')(self);
+    require('./lib/improveCursor.js')(self);
+  },
+  afterConstruct: (self) => {
+    self.pushAssets();
+    self.addRoutes();
+    // self.pushCreateSingleton();
   }
 };
